pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canary-fastrtc"
version = "0.1.0"
description = "NVIDIA Canary-1B-v2 wrapper for FastRTC speech-to-text"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Codeblockz", email = "ryan.a.ellis.1@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.22.0",
    "torch>=2.0.0",
    "nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git@main",
]

[project.urls]
"Homepage" = "https://github.com/Codeblockz/canary-fastrtc"
"Bug Tracker" = "https://github.com/Codeblockz/canary-fastrtc/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]
audio = [
    "librosa>=0.9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
